Move models backend_call.py contents to core/utils.py
